First attempt at variables, including JSON parsing:
given a list of variables like so - ['variable1': 'value1', 'variable2': '{"name": "odd looking string"}']

A simple variable of ${variable1} will be evaluated to "value1".
${variable2.name} will be evaluated to "odd looking string".

This uses the JsonPath library (https://github.com/jayway/JsonPath/blob/master/README.md), which seems pretty good so far. Can't guarantee it will work for the more intricate use-cases we will have in future but will cross that bridge when we get to it.